First stab at 'tetru' but now I need to upgrade GCC to get std::popcount ugh
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@4541 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ pingu/tetris/encode.cc
@@ -656,7 +656,7 @@ static vector<Move> Encode(uint8 target,
 static void PrintSolution(uint8 target,
                           const std::vector<Move> &movie, bool loud) {
   const uint16 full_target = Encoding::FullTarget(target);
-  printf("Done in %lld moves!\n", movie.size());
+  printf("Done in %d moves!\n", (int)movie.size());
 
   // Replay.
   Tetris replay;
